barplot.enrichResult {enrichplot}	R Documentation

barplot

Description

barplot of enrichResult

Usage

## S3 method for class 'enrichResult'
barplot(
  height,
  x = "Count",
  color = "p.adjust",
  showCategory = 8,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)

Arguments

height
enrichResult object

x
one of 'Count' and 'GeneRatio'

color
one of 'pvalue', 'p.adjust' and 'qvalue'

showCategory
number of categories to show

font.size
font size

title
plot title

label_format
a numeric value sets wrap length, alternatively a custom function to format axis labels. by default wraps names longer that 30 characters

...
other parameter, ignored

Value

ggplot object

Examples

library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
barplot(x)
# use `showCategory` to select the displayed terms. It can be a number of a vector of terms.
barplot(x, showCategory = 10)
categorys <- c("pre-malignant neoplasm", "intestinal disease",
               "breast ductal carcinoma", "non-small cell lung carcinoma")
barplot(x, showCategory = categorys)

[Package enrichplot version 1.14.2 Index]