gseaplot2 {enrichplot}	R Documentation

gseaplot2

Description

GSEA plot that mimic the plot generated by broad institute's GSEA software

Usage

gseaplot2(
  x,
  geneSetID,
  title = "",
  color = "green",
  base_size = 11,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  pvalue_table = FALSE,
  ES_geom = "line"
)

Arguments

x
gseaResult object

geneSetID
gene set ID

title
plot title

color
color of running enrichment score line

base_size
base font size

rel_heights
relative heights of subplots

subplots
which subplots to be displayed

pvalue_table
whether add pvalue table

ES_geom
geom for plotting running enrichment score, one of 'line' or 'dot'

Value

plot

Author(s)

Guangchuang Yu

[Package enrichplot version 1.14.2 Index]