ridgeplot {enrichplot}	R Documentation

ridgeplot

Description

ridgeline plot for GSEA result

Usage

ridgeplot(
  x,
  showCategory = 30,
  fill = "p.adjust",
  core_enrichment = TRUE,
  label_format = 30,
  ...
)

## S4 method for signature 'gseaResult'
ridgeplot(
  x,
  showCategory = 30,
  fill = "p.adjust",
  core_enrichment = TRUE,
  label_format = 30,
  ...
)

ridgeplot.gseaResult(
  x,
  showCategory = 30,
  fill = "p.adjust",
  core_enrichment = TRUE,
  label_format = 30,
  orderBy = "NES",
  decreasing = FALSE
)

Arguments

x
gseaResult object

showCategory
number of categories for plotting

fill
one of "pvalue", "p.adjust", "qvalue"

core_enrichment
whether only using core_enriched genes

label_format
a numeric value sets wrap length, alternatively a custom function to format axis labels.

...
additional parameters by default wraps names longer that 30 characters

orderBy
The order of the Y-axis

decreasing
logical. Should the orderBy order be increasing or decreasing?

Value

ggplot object

Author(s)

Guangchuang Yu

Examples

library(DOSE)
data(geneList)
x <- gseDO(geneList)
ridgeplot(x)

[Package enrichplot version 1.14.2 Index]